requirements.txt
django==1.8.5
djangorestframework==3.2.4
python-dateutil==2.4.2

pytest-django==2.9.1
requests==2.8.1
